requirements.txt
# pip install torch==2.1.2+cpu torchaudio==2.1.2+cpu -f https://download.pytorch.org/whl/torch_stable.html
# pip install -r requirements.txt


python-dotenv
openai
requests
PyMySQL
fastapi[standard]
pydantic
pydantic-settings
cryptography
python-jose[cryptography]
sqlmodel
cachetools>=5,<6
pandas
openpyxl

# Audio processing
librosa==0.10.2.post1
numpy>=1.23.0,<2.0
scipy>=1.10.0
soundfile>=0.12.1
noisereduce>=3.0.0
webrtcvad-wheels

# Vector DB & ML
chromadb>=0.4.0
transformers==4.35.2
torch
torchaudio
huggingface_hub>=0.24.0
accelerate>=0.25.0
tqdm>=4.66.0
speechbrain==0.5.14

# LangChain
langchain==0.1.0
langchain-openai>=0.0.5
langchain-core>=0.1.0

apscheduler